denylist set in a fastapi dependency
